requirements.txt
langchain-openai
langchain-community
langchain-core
requests
duckduckgo-search
python-dotenv
